update message state
