+++ src/App.js
@@ -1,8 +1,6 @@
 // import './App.css';
 import React, { useState } from 'react';
 import Account from './components/Account';
-import Login from './components/Login';
-import SignUp from './components/SignUp';
 
 function App() {
   const [isLoggedIn,setIsLoggedIn] = useState(true);
