+++ src/components/SignUp.js
@@ -5,30 +5,47 @@ import {
 //use this link to go to the signUp page 'login'
 
 const SignUp = (props) => {
+    const [message, setMessage] = useState('')
+
+    const updateMessage = msg => {
+        setMessage(msg)
+    }
+
     const [ formData, setFormData] = useState({
         email: '',
         password: '',
         passwordConf: ''
     })
 
-    const handleChange = e => {
-        props.updateMessage('')
-        setFormData({
-            ...formData, 
-            [e.target.name]: e.target.value,
-        })
-    }
-    
+
     const { email, password, passwordConf } = formData
 
     const isFormInvalid = () => {
         return !(email && password && password === passwordConf)
     }
 
+    const handleSubmit = async e => {
+        e.preventDefault()
+        try {
+            // submission logic here
+            // authservice...
+        } catch (err) {
+            updateMessage(err.message)
+        }
+    }
+
+    const handleChange = e => {
+        updateMessage('')
+        setFormData({
+            ...formData, 
+            [e.target.name]: e.target.value,
+        })
+    }
+
     return (
         <form 
             autoComplete='off'
-            onSubmit={props.handleSubmit}
+            onSubmit={handleSubmit}
             className='signupContainer'
         >
             <div>
